+++ src/LibHeifVersion.cs
@@ -57,6 +57,14 @@ namespace LibHeifSharp
         internal static bool Is1Point13OrLater => LibHeifInfo.VersionNumber >= 0x010D0000;
 
         /// <summary>
+        /// Gets a value indicating whether the LibHeif version is at least 1.15.0.
+        /// </summary>
+        /// <value>
+        ///   <c>true</c> if the LibHeif version is at least 1.15.0; otherwise, <c>false</c>.
+        /// </value>
+        internal static bool Is1Point15OrLater => LibHeifInfo.VersionNumber >= 0x010F0000;
+
+        /// <summary>
         /// Throws an exception if the LibHeif version is not supported.
         /// </summary>
         /// <exception cref="HeifException">The LibHeif version is not supported.</exception>